\name{as.task}
\alias{as.task}
\title{Coerce input to a task.}
\usage{
  as.task(x = NULL)
}
\arguments{
  \item{x}{object to coerce to a task}
}
\description{
  Possibile specifications of task: \itemize{ \item
  \code{NULL}: use last task \item path to task \item task
  object }
}
\keyword{internal}

