% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_approve_col.R
\name{create_approve_col}
\alias{create_approve_col}
\title{Update the approve column if keep is in the Answer.surveycode and if the Answer.surverycode is not in the fail vector (failed attention check)}
\usage{
create_approve_col(df, keep, fail)
}
\arguments{
\item{df}{the batch data set provided by amazon}

\item{keep}{the keep character vector that gets returned from the discard_list function}

\item{fail}{character vector of participants that failed the attention check}
}
\value{
a data frame with the approve column manipulated: will have X is valid, empty character if not
}
\description{
Update the approve column if keep is in the Answer.surveycode and if the Answer.surverycode is not in the fail vector (failed attention check)
}
\examples{
clean_batch <- create_approve_col(batch, keep = keep, fail = failed)
}
