% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reject_col.R
\name{create_reject_col}
\alias{create_reject_col}
\title{Update the Reject column based on the approve column}
\usage{
create_reject_col(df, keep, fail)
}
\arguments{
\item{df}{the batch data frame provided from the create_approve_col() function}

\item{keep}{The keep character that gets returned from the discard_list function}

\item{fail}{character vector of participants that failed the attention check}
}
\value{
a data frame with the reject column that has been updated" it will inform that they either failed the attention check or that the mturk code does not match.
If both are not the case then participants will be comnpensated and the reponse coloum willl have an epmty string
}
\description{
Update the Reject column based on the approve column
}
\examples{
create_reject_col(batch_data, keep, fail)
}
