% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_list.R
\name{discard_list}
\alias{discard_list}
\title{Create a list with participant numbers that are both in the batch data set and the mturk data set}
\usage{
discard_list(df1, df2)
}
\arguments{
\item{df1}{the batch data set provided by amazon}

\item{df2}{the mturk data set that we created locally}
}
\value{
a character vector
}
\description{
Create a list with participant numbers that are both in the batch data set and the mturk data set
}
\examples{
discard_list(batch_data, mturk_data)
}
