% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failed_attention_check.R
\name{failed_attention_check}
\alias{failed_attention_check}
\title{Create a vector with all the participants that failed the attention check}
\usage{
failed_attention_check(df, question)
}
\arguments{
\item{df}{data frame with the attention check question}

\item{question}{the attention check question as a string}
}
\value{
a vector of all the participants that failed the attention check
}
\description{
Create a vector with all the participants that failed the attention check
}
\examples{
attention_question <- "This is an attention check"
failed_attention_check <- (petition_data, attention_question)
}
