\name{GenerateQuestionForm}
\alias{GenerateQuestionForm}
\title{Generate QuestionForm}
\description{Generate a QuestionForm data structure. These functions are not yet operational.}
\usage{
GenerateQuestionForm()
}
%\arguments{}
\details{These functions are not yet operational.}
\value{\code{GenerateQuestionForm} returns a character string containing an XML-formatted QuestionForm data object.}
\references{
\href{http://docs.amazonwebservices.com/AWSMechTurk/latest/AWSMturkAPI/ApiReference_QuestionFormDataStructureArticle.html}{API Reference: QuestionForm}
}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{}
\keyword{HITs}
\keyword{Qualifications}