\name{get_balance}
\alias{get_balance}
\title{Retrieves your remaining balance.}
\usage{
  get_balance(task = NULL, ...)
}
\arguments{
  \item{task}{The location of an mturk task, see
  \code{\link{as.task}} for specification options}

  \item{...}{Other parameters passed on to
  \code{\link{mturk_task_req}}.}
}
\description{
  Retrieves your remaining balance.
}

