\name{Use Case: Categorization}
\alias{categorization}
\concept{categorization}
\concept{moderation}
\concept{content moderation}
\concept{photo moderation}
\title{Use Case: Categorization}
\description{This page describes how to use MTurkR to collect categorization data (perhaps for photo moderation or developing a training set) on Amazon Mechanical Turk.}
%\details{}
%\examples{\dontrun{}}
\author{Thomas J. Leeper}
\seealso{
\code{\link{BulkCreate}}

\code{\link{GetAssignments}}
}
\keyword{HITs}
\keyword{Use Cases}
