\name{Use Case: Webscraping}
\alias{webscraping}
\concept{scraping}
\concept{webscraping}
\concept{web scraping}
\title{Use Case: Webscraping}
\description{This page describes how to use MTurkR to scrape human-readable data from the web on Amazon Mechanical Turk.}
%\details{}
%\examples{\dontrun{}}
\author{Thomas J. Leeper}
\seealso{
\code{\link{BulkCreate}}

\code{\link{GetAssignments}}
}
\keyword{HITs}
\keyword{Use Cases}
