\name{Use Case: Sentiment Analysis}
\alias{sentiment}
\concept{sentiment}
\concept{sentiment analysis}
\title{Use Case: Sentiment Analysis}
\description{This page describes how to use MTurkR to collect sentiment data on Amazon Mechanical Turk.}
%\details{}
%\examples{\dontrun{}}
\author{Thomas J. Leeper}
\seealso{
\code{\link{BulkCreate}}

\code{\link{GetAssignments}}
}
\keyword{HITs}
\keyword{Use Cases}
