\name{Use Case: Surveys}
\alias{survey}
\concept{survey}
\concept{survey survey}
\concept{survey panel}
\concept{questionnaire}
\title{Use Case: Surveys}
\description{This page describes how to use MTurkR to collect survey(-experimental) data on Amazon Mechanical Turk.}
%\details{}
%\examples{\dontrun{}}
\author{Thomas J. Leeper}
\seealso{
\code{\link{CreateHIT}}

\code{\link{GetAssignments}}

\code{\link{ContactWorkers}}

\code{\link{GrantBonus}}
}
\keyword{HITs}
\keyword{Use Cases}
