\name{GenerateAnswerKey}
\alias{GenerateAnswerKey}
\title{Generate AnswerKey}
\description{Generate an AnswerKey data structure for a Qualification test. This is not currently functional}
\usage{
GenerateAnswerKey()
}
\details{This is not yet operational. An AnswerKey serves to automatically score a Qualification test, as specified in the \code{test} parameter of \code{\link{CreateQualificationType}}.}
\value{A character string containing an XML-formatted AnswerKey data structure.}
\references{
\href{http://docs.amazonwebservices.com/AWSMechTurk/latest/AWSMturkAPI/ApiReference_AnswerKeyDataStructureArticle.html}{API Reference}
}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{}
\keyword{HITs}