\name{Use Case: Webscraping}
\alias{webscraping}
\concept{scraping}
\concept{webscraping}
\concept{web scraping}
\title{Use Case: Webscraping}
\description{This page describes how to use MTurkR to scrape human-readable data from the web on Amazon Mechanical Turk.}
%\details{}
%\examples{\dontrun{}}
\author{Thomas J. Leeper}
\seealso{
For guidance on some of the functions used in this tutorial, see:
\itemize {
    \item \code{\link{BulkCreate}}
    \item \code{\link{HITStatus}}
    \item \code{\link{GetAssignments}}
}

For some other tutorials on how to use MTurkR for specific use cases, see the following:
\itemize{
    \item \link{survey}, for collecting survey(-experimental) data
    \item \link{categorization}, for doing large-scale categorization (e.g., photo moderation or developing a training set)
    \item \link{sentiment}, for doing sentiment coding
    %\item \link{webscraping}, for manual scraping of web data

}
\keyword{Use Cases}
